Delete polling soft purge job.
